fix: improved script to build Croupier wheel

+++ release_croupier.sh
@@ -1,5 +1,12 @@
+#!/bin/bash
+echo $# arguments
+if [ "$#" -ne 1 ]; then
+    echo "Error: croupier branch not provided. Use: release_croupier <branch>" >&2
+    exit 2
+fi
+
 rm croupier.zip
 rm *.wgn
-git pull
+git pull origin $1
 zip -qr croupier.zip ../croupier
 wagon create -f ./croupier.zip -a '--no-cache-dir -c ./constraints.txt'
